Remove Groovy dependency from network store client by moving tools to another Maven module
