+++ network-store-tools/src/main/java/com/powsybl/network/store/tools/NetworkStoreListTool.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.network.store.client.tools;
+package com.powsybl.network.store.tools;
 
 import com.google.auto.service.AutoService;
 import com.powsybl.network.store.client.NetworkStoreConfig;
